+++ tests/test_eval.py
@@ -6,8 +6,8 @@ from typing import List
 import pytest
 from pytest import LogCaptureFixture
 
+from commands.eval.classes.helpers import RunSettings
 from commands.eval.classes.run_object import RunObject
-from commands.eval.classes.run_settings import RunSettings
 from commands.eval.main import main
